Finalized Project

+++ src/main/LCD.h
@@ -16,7 +16,7 @@ public:
   void clear();
 
 private:
-  int waterLevel;
+  float waterLevel;
   String state;  
   bool isON;
   LiquidCrystal_I2C lcd = LiquidCrystal_I2C(0x27,20,4);
